Recommend test-framework-quickcheck2
The doc currently refers to `test-framework-quickcheck`, which doesn't support the latest version of QuickCheck. After a bit of poking around, it seems that `test-framework-quickcheck2` does, though. I'm a bit confused on why there are two of them, but it seems like the second one is the one that we ought to be recommending.